# swagger_output.json
{
  "swagger": "2.0",
  "info": {
    "version": "1.0.0",
    "title": "API E-PARK",
    "description": "O E-Park é um aplicativo desenvolvido para facilitar a reserva e marcação de vagas em estacionamentos privativos. Os usuários podem encontrar vagas disponíveis, realizar o cadastro de seus veículos, efetuar o pagamento através de cartão, e receber o ticket de confirmação da reserva através do aplicativo onde a aplicação está diretamente relacionada com o gerenciador de sistema E-park web."
  },
  "host": "localhost:3000",
  "basePath": "/",
  "schemes": [
    "http"
  ],
  "paths": {
    "/Usuario/": {
      "get": {
        "description": "",
        "tags": ["Usuario"],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "description": "",
        "tags": ["Usuario"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "type": "string",
                  "example": "any"
                },
                "email": {
                  "type": "string",
                  "example": "any"
                },
                "senha": {
                  "type": "string",
                  "example": "any"
                },
                "foto": {
                  "type": "string",
                  "example": "any"
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "Created"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Usuario/{id}": {
      "get": {
        "description": "",
        "tags": ["Usuario"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "usuário encontrado"
          },
          "404": {
            "description": "usuário não encontrado"
          }
        }
      },
      "put": {
        "description": "Atualiza um usuário pelo ID",
        "tags": ["Usuario"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "nome": {
                  "type": "string",
                  "example": "any"
                },
                "email": {
                  "type": "string",
                  "example": "any"
                },
                "senha": {
                  "type": "string",
                  "example": "any"
                },
                "foto": {
                  "type": "string",
                  "example": "any"
                }
              },
              "required": ["nome", "email", "senha"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "404": {
            "description": "Usuário não encontrado"
          }
        }
      },
      "delete": {
        "description": "",
        "tags": ["Usuario"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          },
          "500": {
            "description": "Internal Server Error"
          }
        }
      }
    },
    "/Veiculo/": {
      "get": {
        "description": "",
        "tags": ["Veiculo"],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "description": "",
        "tags": ["Veiculo"],
        "parameters": [
          {
            "name": "body",
            "in": "body",
            "schema": {
              "type": "object",
              "properties": {
                "usuarioid": {
                  "type": "string",
                  "example": ""
                },
                "marca": {
                  "type": "string",
                  "example": ""
                },
                "modelo": {
                  "type": "string",
                  "example": ""
                },
                "ano": {
                  "type": "integer",
                  "example": 0
                },
                "placa": {
                  "type": "string",
                  "example": ""
                }
              }
            }
          }
        ],
        "responses": {
          "201": {
            "description": "veículo criado"
          },
          "400": {
            "description": "usuarioid é obrigatório"
          },
          "404": {
            "description": "usuarioid não encontrado"
          }
        }
      }
    },
    "/Veiculo/{id}": {
      "get": {
        "description": "",
        "tags": ["Veiculo"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }, 
          "500": {
            "description": "Something went wrong!"
          },
          "404": {
            "description": "Veiculo não encontrado"
          }
        }
      },
      "put": {
        "description": "Atualiza um veículo pelo ID",
        "tags": ["Veiculo"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          },
          {
            "name": "body",
            "in": "body",
            "required": true,
            "schema": {
              "type": "object",
              "properties": {
                "usuarioid": {
                  "type": "string"
                },
                "marca": {
                  "type": "string",
                  "example": "any"
                },
                "modelo": {
                  "type": "string",
                  "example": "any"
                },
                "ano": {
                  "type": "integer",
                  "example": "any"
                },
                "placa": {
                  "type": "string",
                  "example": "any"
                }
              },
              "required": ["usuarioid", "marca", "modelo", "ano", "placa"]
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Veículo atualizado com sucesso"
          },
          "404": {
            "description": "Veículo não encontrado"
          },
          "500": {
            "description": "Erro ao atualizar veículo"
          }
        }
      },
      "delete": {
        "description": "",
        "tags": ["Veiculo"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Vaga/": {
      "get": {
        "description": "",
        "tags": ["Vaga"],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "description": "",
        "tags": ["Vaga"],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Vaga/{id}": {
      "get": {
        "description": "",
        "tags": ["Vaga"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "description": "",
        "tags": ["Vaga"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "description": "",
        "tags": ["Vaga"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Vagas/ocupadas": {
      "get": {
        "summary": "Retorna o número de vagas ocupadas, separadas por preferencial e normal",
        "tags": ["Vaga"],
        "responses": {
          "200": {
            "description": "Número de vagas ocupadas retornado com sucesso",
            "schema": {
              "type": "object",
              "properties": {
                "total_preferencial": {
                  "type": "integer",
                  "description": "Número total de vagas preferenciais ocupadas"
                },
                "total_normal": {
                  "type": "integer",
                  "description": "Número total de vagas normais ocupadas"
                }
              }
            }
          }
        }
      }
    },
    "/Reserva/": {
      "get": {
        "description": "",
        "tags": ["Reserva"],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "description": "",
        "tags": ["Reserva"],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Reserva/{id}": {
      "get": {
        "description": "",
        "tags": ["Reserva"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "description": "",
        "tags": ["Reserva"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "description": "",
        "tags": ["Reserva"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Ticket/": {
      "get": {
        "description": "",
        "tags": ["Ticket"],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "post": {
        "description": "",
        "tags": ["Ticket"],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/Ticket/{id}": {
      "get": {
        "description": "",
        "tags": ["Ticket"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "put": {
        "description": "",
        "tags": ["Ticket"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      },
      "delete": {
        "description": "",
        "tags": ["Ticket"],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "type": "string"
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    }
  }
}